feat(api): vector-output 1D fits and caller-owned batch Scratch
Support array<T,1>->array<T,K> fits and accept array-input/vector-output
fits in the 1D batch (un)sorted overloads. Add Function::Scratch, a
caller-owned, allocator-aware batch scratch with uninitialised
allocations and a make_scratch factory. Recompute the leaf id during
scatter (dropping the leaf_ids buffer) with an AVX-512DQ vcvttpd2qq
quantize, add a min_uniform_depth knob and leaf-table state in
print_stats, and gate -Wno-c2y-extensions on clang>=19.

+++ examples/c++/baobzi_bench_pack_scatter.cpp
@@ -181,7 +181,7 @@ void sweep_scatter_1d(ankerl::nanobench::Bench &b, const char *label,
     std::vector<const fit_t *> fit_ptrs;
     for (auto &fn : fits) fit_ptrs.push_back(&fn);
 
-    std::mt19937 ig(13 + R);
+    std::mt19937 ig(static_cast<std::uint32_t>(13 + R));
     std::uniform_int_distribution<std::uint32_t> rd(
         0, static_cast<std::uint32_t>(R - 1));
     std::uniform_real_distribution<double> td(1e-6, 1.0 - 1e-6);
